\name{formats}
\alias{formats}
\alias{io}
\alias{datasets}
\alias{methods}
\alias{dag_ml}
\alias{dag_ml_data}
\title{Access upstream nirs4all domains}
\usage{
formats()

io()

datasets()

methods()

dag_ml()

dag_ml_data()
}
\value{
The loaded namespace for the requested upstream domain.
}
\description{
Convenience accessors for the upstream domains exposed by the aggregate
\pkg{nirs4all} R package.
}
